+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+0.2.0 (2021-08-14)
+------------------
+
+(WIP -- PENDING)
+
+* Fixes authentication (thanks @Lustyn)
+* Replaces setuptools with filt
+* Replaces click with typer
+* Replaces requests with httpx
+
 0.1.0 (2018-12-25)
 ------------------